requirements.txt
tensorflow==2.13.0
opencv-python==4.8.0.74
numpy==1.24.3
Pillow==10.0.0